+++ src/command_modules/azure-cli-appservice/azure/cli/command_modules/appservice/_params.py
@@ -28,14 +28,25 @@ two_states_switch = ['true', 'false']
 def _generic_site_operation(resource_group_name, name, operation_name, slot=None,
                             extra_parameter=None, client=None):
     client = client or web_client_factory()
-    m = getattr(client.web_apps,
-                operation_name if slot is None else operation_name + '_slot')
+    operation = getattr(client.web_apps,
+                        operation_name if slot is None else operation_name + '_slot')
     if slot is None:
-        return (m(resource_group_name, name)
-                if extra_parameter is None else m(resource_group_name, name, extra_parameter))
+        return (operation(resource_group_name, name)
+                if extra_parameter is None else operation(resource_group_name,
+                                                          name, extra_parameter))
 
-    return (m(resource_group_name, name, slot)
-            if extra_parameter is None else m(resource_group_name, name, extra_parameter, slot))
+    return (operation(resource_group_name, name, slot)
+            if extra_parameter is None else operation(resource_group_name,
+                                                      name, extra_parameter, slot))
+
+
+def _generic_settings_operation(resource_group_name, name, operation_name, setting_properties, slot=None, client=None):
+    client = client or web_client_factory()
+    m = getattr(client.web_apps, operation_name if slot is None else operation_name + '_slot')
+    if slot is None:
+        return m(resource_group_name, name, str, setting_properties)
+
+    return m(resource_group_name, name, slot, str, setting_properties)
 
 
 def get_hostname_completion_list(prefix, action, parsed_args, **kwargs):  # pylint: disable=unused-argument